SelectionTool : Register `Shader Source` selector
